+++ typescript/src/optimistic-minting.ts
@@ -1,103 +1,6 @@
-import { BigNumber } from "ethers"
 import { TransactionHash } from "./lib/bitcoin"
-import { Identifier, Event, TBTCVault } from "./lib/contracts"
+import { TBTCVault, OptimisticMintingRequest } from "./lib/contracts"
 import { Hex } from "./lib/utils"
-
-/**
- * Represents an event that is emitted when a new optimistic minting is requested
- * on chain.
- */
-export type OptimisticMintingRequestedEvent = {
-  /**
-   * Minter's chain identifier.
-   */
-  minter: Identifier
-  /**
-   * Unique deposit identifier.
-   * @see Bridge.buildDepositKey
-   */
-  depositKey: Hex
-  /**
-   * Depositor's chain identifier.
-   */
-  depositor: Identifier
-  /**
-   * Amount of tokens requested to mint.
-   * This value is in ERC 1e18 precision, it has to be converted before using
-   * as Bitcoin value with 1e8 precision in satoshi.
-   */
-  // TODO: Consider adding a custom type to handle conversion from ERC with 1e18
-  //       precision to Bitcoin in 1e8 precision (satoshi).
-  amount: BigNumber
-  /**
-   * Hash of a Bitcoin transaction made to fund the deposit.
-   */
-  fundingTxHash: TransactionHash
-  /**
-   * Index of an output in the funding transaction made to fund the deposit.
-   */
-  fundingOutputIndex: number
-} & Event
-
-/**
- * Represents an event that is emitted when an optimistic minting request
- * is cancelled on chain.
- */
-export type OptimisticMintingCancelledEvent = {
-  /**
-   * Guardian's chain identifier.
-   */
-  guardian: Identifier
-  /**
-   * Unique deposit identifier.
-   * @see Bridge.buildDepositKey
-   */
-  depositKey: Hex
-} & Event
-
-/**
- * Represents an event that is emitted when an optimistic minting request
- * is finalized on chain.
- */
-export type OptimisticMintingFinalizedEvent = {
-  /**
-   * Minter's chain identifier.
-   */
-  minter: Identifier
-  /**
-   * Unique deposit identifier.
-   * @see Bridge.buildDepositKey
-   */
-  depositKey: Hex
-  /**
-   * Depositor's chain identifier.
-   */
-  depositor: Identifier
-  /**
-   * Value of the new optimistic minting debt of the depositor.
-   * This value is in ERC 1e18 precision, it has to be converted before using
-   * as Bitcoin value with 1e8 precision in satoshi.
-   */
-  // TODO: Consider adding a custom type to handle conversion from ERC with 1e18
-  //       precision to Bitcoin in 1e8 precision (satoshi).
-  optimisticMintingDebt: BigNumber
-} & Event
-
-/**
- * Represents optimistic minting request for the given deposit revealed to the
- * Bridge.
- */
-export type OptimisticMintingRequest = {
-  /**
-   * UNIX timestamp at which the optimistic minting was requested.
-   */
-  requestedAt: number
-  /**
-   * UNIX timestamp at which the optimistic minting was finalized.
-   * 0 if not yet finalized.
-   */
-  finalizedAt: number
-}
 
 /**
  * Requests optimistic minting for a deposit on chain.
